Parse the existing config file
